+++ src/main/java/com/fot/backcontask/service/company/CompanyService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.company;
 import com.fot.backcontask.model.Company;
 import com.fot.backcontask.service.Service;
 
-public interface CompanyService extends Service<Company, Integer> {
+public interface CompanyService extends Service<Company, Long> {
 
 }
