+++ src/main/java/com/fot/backcontask/service/companytype/CompanyTypeService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.companytype;
 import com.fot.backcontask.model.CompanyType;
 import com.fot.backcontask.service.Service;
 
-public interface CompanyTypeService extends Service<CompanyType, Integer> {
+public interface CompanyTypeService extends Service<CompanyType, Long> {
 
 }
