+++ src/main/java/com/fot/backcontask/service/contractline/ContractLineService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.contractline;
 import com.fot.backcontask.model.ContractLine;
 import com.fot.backcontask.service.Service;
 
-public interface ContractLineService extends Service<ContractLine, Integer> {
+public interface ContractLineService extends Service<ContractLine, Long> {
 
 }
