+++ src/main/java/com/fot/backcontask/service/privilege/PrivilegeService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.privilege;
 import com.fot.backcontask.model.Privilege;
 import com.fot.backcontask.service.Service;
 
-public interface PrivilegeService extends Service<Privilege, Integer> {
+public interface PrivilegeService extends Service<Privilege, Long> {
 
 }
