+++ src/main/java/com/fot/backcontask/service/role/RoleService.java
@@ -3,6 +3,6 @@ package com.fot.backcontask.service.role;
 import com.fot.backcontask.model.Role;
 import com.fot.backcontask.service.Service;
 
-public interface RoleService extends Service<Role, Integer> {
+public interface RoleService extends Service<Role, Long> {
 
 }
